Change template strings for a template language
Previously we were using template literals for generating the bot
responses but when you need to do a loop, that's a bit odd so this PR
switches to Handlebars to build the text output.

It pre compiles the templates when the module loads and then consumes
the compiled templates as a function.